fixed number bug
updated lexer to use doubles

+++ util/val.cpp
@@ -1,11 +1,12 @@
 #include "val.hpp"
 
+
 Val::Val(std::string val) {
     type = String;
     string = val;
 }
 
-Val::Val(float val) {
+Val::Val(double val) {
     type = Number;
     number = val;
 }
